% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvseries.R
\docType{package}
\name{tvseries}
\alias{tvseries}
\alias{tvseries-package}
\title{IMBDb ratings of TV series}
\format{A dataframe with 3,627,538 rows and 10 columns.
\itemize{
    \item \strong{tvseries_title}: Name of the TV series.
    \item \strong{episode_title}: Name of the episode.
    \item \strong{season_number}: Season number.
    \item \strong{episode_number}: Episode number.
    \item \strong{start_year}: Year the show started.
    \item \strong{end_year}: Year the show ended.
    \item \strong{runtime_min}: Length of the episode in minutes.
    \item \strong{genres}: Genres of the TV series.
    \item \strong{average_rating}: Average rating of the episode.
    \item \strong{votes}: Number of votes for that episode. }}
\source{
\url{https://datasets.imdbws.com/}
}
\usage{
tvseries
}
\description{
This package contains ratings an other TV series information from
IMDb.

A dataset containing IMDb ratings for all series, seasons and
episodes. Be careful when loading this dataset as it is pretty
big and might freeze your session.
}
\keyword{datasets}
