% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvseries.R
\docType{data}
\name{tvseries_top100}
\alias{tvseries_top100}
\title{Top 100 most popular IMDB-rated TV series}
\format{A dataframe with 7,100 rows and 10 columns.
\itemize{
    \item \strong{tvseries_title}: Name of the TV series.
    \item \strong{episode_title}: Name of the episode.
    \item \strong{season_number}: Season number.
    \item \strong{episode_number}: Episode number.
    \item \strong{start_year}: Year the show started.
    \item \strong{end_year}: Year the show ended.
    \item \strong{runtime_min}: Length of the episode in minutes.
    \item \strong{genres}: Genres of the TV series.
    \item \strong{average_rating}: Average rating of the episode.
    \item \strong{votes}: Number of votes for that episode. }}
\source{
\url{https://datasets.imdbws.com/}
}
\usage{
tvseries_top100
}
\description{
A dataset containing IMDb ratings for the top 100 TV series selected
by their popularity (average number of votes per episode) and their
average ranking.
}
\keyword{datasets}
